Refactor and prepare rpq experiment
