+++ project/fa_utils.py
@@ -4,7 +4,7 @@ import networkx as nx
 import pyformlang.finite_automaton
 import pyformlang.regular_expression
 
-from project.pygraphblas_bool_decomposed_nfa import BoolDecomposedNFA
+from project.scipy_sparse_bool_decomposed_nfa import BoolDecomposedNFA
 
 __all__ = ["graph_to_nfa", "regex_to_minimal_dfa", "intersect_nfas"]
